break out enqueue channel
