+++ client/session.go
@@ -1976,148 +1976,6 @@ type receivedBlocks struct {
 
 type processFn func(batch []*blocksMetadata)
 
-type peerBlocksQueue struct {
-	sync.RWMutex
-	closed       bool
-	peer         hostQueue
-	queue        []*blocksMetadata
-	doneFns      []func()
-	assigned     uint64
-	completed    uint64
-	maxQueueSize int
-	workers      xsync.WorkerPool
-	processFn    processFn
-}
-
-type newPeerBlocksQueueFn func(
-	peer hostQueue,
-	maxQueueSize int,
-	interval time.Duration,
-	workers xsync.WorkerPool,
-	processFn processFn,
-) *peerBlocksQueue
-
-func newPeerBlocksQueue(
-	peer hostQueue,
-	maxQueueSize int,
-	interval time.Duration,
-	workers xsync.WorkerPool,
-	processFn processFn,
-) *peerBlocksQueue {
-	q := &peerBlocksQueue{
-		peer:         peer,
-		maxQueueSize: maxQueueSize,
-		workers:      workers,
-		processFn:    processFn,
-	}
-	if interval > 0 {
-		go q.drainEvery(interval)
-	}
-	return q
-}
-
-func (q *peerBlocksQueue) drainEvery(interval time.Duration) {
-	for {
-		q.Lock()
-		if q.closed {
-			q.Unlock()
-			return
-		}
-		q.drainWithLock()
-		q.Unlock()
-		time.Sleep(interval)
-	}
-}
-
-func (q *peerBlocksQueue) close() {
-	q.Lock()
-	q.closed = true
-	q.Unlock()
-}
-
-func (q *peerBlocksQueue) trackAssigned(amount int) {
-	atomic.AddUint64(&q.assigned, uint64(amount))
-}
-
-func (q *peerBlocksQueue) trackCompleted(amount int) {
-	atomic.AddUint64(&q.completed, uint64(amount))
-}
-
-func (q *peerBlocksQueue) enqueue(bm *blocksMetadata, doneFn func()) {
-	q.Lock()
-
-	if len(q.queue) == 0 && cap(q.queue) < q.maxQueueSize {
-		// Lazy initialize queue
-		q.queue = make([]*blocksMetadata, 0, q.maxQueueSize)
-	}
-	if len(q.doneFns) == 0 && cap(q.doneFns) < q.maxQueueSize {
-		// Lazy initialize doneFns
-		q.doneFns = make([]func(), 0, q.maxQueueSize)
-	}
-	q.queue = append(q.queue, bm)
-	if doneFn != nil {
-		q.doneFns = append(q.doneFns, doneFn)
-	}
-	q.trackAssigned(len(bm.blocks))
-
-	// Determine if should drain immediately
-	if len(q.queue) < q.maxQueueSize {
-		// Require more to fill up block
-		q.Unlock()
-		return
-	}
-	q.drainWithLock()
-
-	q.Unlock()
-}
-
-func (q *peerBlocksQueue) drain() {
-	q.Lock()
-	q.drainWithLock()
-	q.Unlock()
-}
-
-func (q *peerBlocksQueue) drainWithLock() {
-	if len(q.queue) == 0 {
-		// None to drain
-		return
-	}
-	enqueued := q.queue
-	doneFns := q.doneFns
-	q.queue = nil
-	q.doneFns = nil
-	q.workers.Go(func() {
-		q.processFn(enqueued)
-		// Call done callbacks
-		for i := range doneFns {
-			doneFns[i]()
-		}
-		// Track completed blocks
-		completed := 0
-		for i := range enqueued {
-			completed += len(enqueued[i].blocks)
-		}
-		q.trackCompleted(completed)
-	})
-}
-
-type peerBlocksQueues []*peerBlocksQueue
-
-func (qs peerBlocksQueues) findQueue(peer hostQueue) *peerBlocksQueue {
-	for _, q := range qs {
-		if q.peer == peer {
-			return q
-		}
-	}
-	return nil
-}
-
-func (qs peerBlocksQueues) closeAll() {
-	for _, q := range qs {
-		q.close()
-	}
-}
-
 type blocksMetadata struct {
 	peer   hostQueue
 	id     ts.ID
